requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
gunicorn==21.2.0
# psycopg2-binary Python 3.13 bilan Renderda xato berishi mumkin; psycopg 3 ishlatamiz
psycopg[binary]==3.2.4
requests==2.31.0
supabase>=2.3.0

